:sparkles: ディレクトリ名の変更
