--- python/hello_world_2/run.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-exec ./main.py World
