+++ python/hello_world_3/run.sh
@@ -1,4 +1,3 @@
 #!/usr/bin/env bash
 
-# Python 3.11じゃないと動かないんだよね……
-exec python3.11 main.py
+exec ./main.py World
