+++ python/hello_world_4/run.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+# Python 3.11じゃないと動かないんだよね……
+exec python3.11 main.py
